requirements.txt
discord.py==2.5.2
openai==1.99.9
httpx==0.27.2
